+++ packages/vite-bundler/build.js
@@ -3,7 +3,7 @@ import { performance } from 'node:perf_hooks'
 import fs from 'fs-extra'
 import { execaSync } from 'execa'
 import pc from 'picocolors'
-import { createWorkerFork, cwd, getProjectPackageJson, getTempDir } from './workers';
+import { createWorkerFork, cwd, getProjectPackageJson, getTempDir, meteorPackagePath } from './workers';
 import EntryFile from './build/EntryFile';
 import BuildResult from './build/BuildResult';
 
@@ -149,6 +149,7 @@ try {
         meteor: {
           packagePath: path.join(tempMeteorOutDir, 'bundle', 'programs', 'web.browser', 'packages'),
           isopackPath: path.join(tempMeteorProject, '.meteor', 'local', 'isopacks'),
+          globalMeteorPackagesDir: meteorPackagePath,
         },
       }],
     })